Refactor: remove deprecated import shim modules
Remove backward compatibility shim modules that were providing deprecated
import paths during the refactoring process. All modules now require
direct imports from their canonical locations.

Removed modules:
- ClassicLib.FileIOCore → Use ClassicLib.FileIO
- ClassicLib.AsyncUtil → Use ClassicLib.FileIO.Async
- ClassicLib.AsyncYamlSettingsCore → Use ClassicLib.YamlSettings
- ClassicLib.AsyncYamlSettings → Use ClassicLib.YamlSettings
- ClassicLib.AsyncUtilities → Use ClassicLib.Utils.Async
- ClassicLib.AsyncUtilities_Enhanced → Use ClassicLib.Utils.Async

Also removed:
- rust_loader.get_rust_module() deprecated function
- MessageTarget.GUI_ONLY and CLI_ONLY deprecated enum values

Updated internal imports and tests to use canonical paths.
